Workaround v8 bugs with small stacks
v8 seems to have some issues when you try and use stack sizes smaller
than it's used to.

First is that if you attempt to set a stack limit that's smaller than
the snapshot size (if enabled) it will crash your shit in a very hard to
understand way. Unfortunately there's no way to actually know what that
magic size is in advance, so I'm just using a secret v8 internal
variable. This means that for most systems each coroutine will consume
~320kb of memory.

Second issue is there's a bug on 64 bit platforms when your Javascript
overflows the stack. It's probably pretty hard to hit this bug in
regular code but it's still quite bad. There's a workaround that I'm
using to get past that.